fix a bug for visitor exit

+++ service/payment/src/main/java/ca/mcmaster/cas735/acmepark/payment/dto/PaymentRequest.java
@@ -19,7 +19,7 @@ public class PaymentRequest {
 
     private String licensePlate;
 
-    private int lotId;
+    private Long lotId;
 
     private int userId;
